requirements.txt
backtrader==1.9.78.123
openai>=1.0.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
pydantic>=2.0.0
aiohttp>=3.8.0
asyncio
typing-extensions 
finnhub-python